New layout written with fomantic ui instead of custom, publish endpoint started, restructured gateway, switched to rerun
